pyproject.toml
[project]
name = "bubbles"
version = "0.1.0"
description = "Project 'bubbles' dependencies"
requires-python = ">=3.8"
dependencies = [
    "Fiona",
    "Rtree",
    "requests",
    "matplotlib",
    "pyproj",
    "facebook-business",
    "python-dotenv",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
